Migrate from web API to CLI
Explanation
===========

Motivation
----------
The previous version of this tool used the GitHub web API directly,
which required prompting for the user's credentials and authenticating
as the user before doing the import.

Given that GitHub now has a command-line tool `gh` which has its own
secure storage for GitHub credentials and tokens, we no longer need to
receive a user's credentials directly in this tool. Instead we'll let
the `gh` CLI tool manage credentials and we'll call the APIs via that
tool.

Implementation
--------------
Swap out all uses of the web API for the CLI tool. The API endpoints are
the same, we're just fronting the API calls with the CLI tool for
credential management.

Risks
-----
This does mean we're limited in terms of which APIs to hit to just the
ones that the CLI implements. So far we don't use any that aren't
implemented in the CLI tool.

Testing
=======
Manually tested these endponts by importing issues.