Skip unnecessary prompts about Jira project
Explanation
===========

Motivation
----------
The previous version of the import prompted for the Jira project name
and the field that indicates an issue is done. Both of these come from
the Jira XML export directly, so there's no value in prompting for them.
If users want to filter the issues by project or status then they
should do then when exporting the Jira issues to XML (just like for any
other attribute they want to filter on).

Implementation
--------------
1. Drop the prompt and filter on Jira project name and just use the name
   from the XML file.
2. Drop the prompt and usage of the Jira status field and just check
   the export for the closed time to determine status.

Risks
-----
This does add an extra step / filter to the Jira export, but that's more
consistent and flexible and easier to use.

Testing
=======
Tested by importing Jira issues.