Allow default labels to be added to all imported issues
Explanation
===========
So that it's easier to manage the imported issues, this adds the ability
to label all imported issues with a specific label or set of labels. It
also fixes handling of some Jira properties like issue type so that they
get imported as labels.

Testing
=======
Tested by importing Jira issues to GitHub and using this feature.