Post-process issue comments and links
Explanation
===========

Motivation
----------
Jira users aren't mapped to GitHub users, but they also weren't cited
correctly, and often they were just showing up as user IDs instead of
names, which is less usable in the comment history. This fixes that so
that comments and issues are properly attributed.

Also, links and attachments in Jira were not getting copied over, so any
relative links in the Jira text were just broken on GitHub import. This
adds the links and images and attachments to the post-processing,
replacing all with absolute links so that they still work.

Implementation
--------------
Post-processes the imported issues:
1. Tracks the names for all referenced users and updates their name and
   link to their Jira profile to replace the user ID.
2. Replaces all relative links and images with absolute links to the
   Jira files.

Risks
-----
This isn't the ideal way to handle either users or attachments, but
hopefully it's a step in the right direction because at least the data
isn't broken. It does still depend on the original Jira source for
linking both users and attachments, so that's probably not ideal for all
use cases where users might be trying to shut down a Jira stack rather
than keep it running for these links.

Testing
=======
Tested by importing more Jira issues to GitHub.